+++ src/features/metricsCodeLensProvider.ts
@@ -10,7 +10,9 @@ import {
   EventEmitter,
   Event,
   workspace,
-  Disposable
+  Disposable,
+  FileSystemWatcher,
+  RelativePattern
 } from "vscode";
 import * as path from "path";
 import * as fs from "fs";
@@ -27,6 +29,7 @@ export class LogtalkMetricsCodeLensProvider implements CodeLensProvider {
 
   private configurationListener: Disposable;
   private textDocumentListener: Disposable;
+  private resultsFileWatchers: FileSystemWatcher[] = [];
 
   constructor() {
     this.configurationListener = workspace.onDidChangeConfiguration((_) => {
@@ -60,6 +63,38 @@ export class LogtalkMetricsCodeLensProvider implements CodeLensProvider {
       LogtalkMetricsCodeLensProvider.outdated = true;
       this._onDidChangeCodeLenses.fire();
     });
+
+    // Watch for changes to metrics results files to refresh CodeLens automatically
+    this.setupResultsFileWatchers();
+  }
+
+  /**
+   * Set up file system watchers for .vscode_metrics_results files in all workspace folders
+   */
+  private setupResultsFileWatchers(): void {
+    // Create watchers for each workspace folder
+    if (workspace.workspaceFolders) {
+      for (const folder of workspace.workspaceFolders) {
+        const pattern = new RelativePattern(folder, '**/.vscode_metrics_results');
+        const watcher = workspace.createFileSystemWatcher(pattern);
+
+        watcher.onDidCreate(() => {
+          LogtalkMetricsCodeLensProvider.outdated = false;
+          this._onDidChangeCodeLenses.fire();
+        });
+
+        watcher.onDidChange(() => {
+          LogtalkMetricsCodeLensProvider.outdated = false;
+          this._onDidChangeCodeLenses.fire();
+        });
+
+        watcher.onDidDelete(() => {
+          this._onDidChangeCodeLenses.fire();
+        });
+
+        this.resultsFileWatchers.push(watcher);
+      }
+    }
   }
 
   /**
@@ -72,6 +107,10 @@ export class LogtalkMetricsCodeLensProvider implements CodeLensProvider {
     if (this.textDocumentListener) {
       this.textDocumentListener.dispose();
     }
+    for (const watcher of this.resultsFileWatchers) {
+      watcher.dispose();
+    }
+    this.resultsFileWatchers = [];
     if (this._onDidChangeCodeLenses) {
       this._onDidChangeCodeLenses.dispose();
     }